requirements.txt
click
numpy
scipy
scikit-image
tqdm
opencv-python
dataclasses
pillow
matplotlib
shapely
ocrd>=1.0.0b11
ocr4all-page-content-detection
